+++ src/main/java/org/kakaoshare/backend/domain/order/dto/inquiry/response/OrderHistoryDetailResponse.java
@@ -1,5 +1,6 @@
-package org.kakaoshare.backend.domain.order.dto.inquiry;
+package org.kakaoshare.backend.domain.order.dto.inquiry.response;
 
+import org.kakaoshare.backend.domain.order.dto.inquiry.OrderHistoryDetailDto;
 import org.kakaoshare.backend.domain.payment.dto.inquiry.PaymentHistoryDto;
 
 public record OrderHistoryDetailResponse(OrderHistoryDetailDto order, PaymentHistoryDto payment) {
